chickfit-ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow>=2.10.0,<2.16.0
pillow==10.0.1
numpy>=1.21.0,<2.0.0
google-cloud-storage==2.10.0
python-multipart==0.0.6
pydantic==2.4.2
python-jose[cryptography]==3.3.0
scikit-learn>=1.0.0,<2.0.0
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0